+++ src/backend/plotters.rs
@@ -10,11 +10,11 @@ use autd3_driver::{autd3_device::AUTD3, geometry::Geometry};
 #[derive(Clone, Debug)]
 pub struct PlotConfig {
     pub figsize: (u32, u32),
-    pub cbar_size: f64,
+    pub cbar_size: f32,
     pub font_size: u32,
     pub label_area_size: u32,
     pub margin: u32,
-    pub ticks_step: f64,
+    pub ticks_step: f32,
     pub cmap: ListedColorMap,
     pub fname: OsString,
 }
@@ -47,13 +47,12 @@ impl PartialEq for PlotConfig {
     }
 }
 
-
 pub struct PlottersBackend {}
 
 impl PlottersBackend {
     fn plot_modulation_impl<B: plotters::backend::DrawingBackend>(
         root: DrawingArea<B, Shift>,
-        modulation: Vec<f64>,
+        modulation: Vec<f32>,
         config: &PlotConfig,
     ) -> Result<(), crate::error::VisualizerError>
     where
@@ -66,7 +65,7 @@ impl PlottersBackend {
             .margin(config.margin)
             .x_label_area_size(config.label_area_size)
             .y_label_area_size(config.label_area_size)
-            .build_cartesian_2d::<_, std::ops::Range<f64>>(0..modulation.len(), 0.0..1.0)?;
+            .build_cartesian_2d::<_, std::ops::Range<f32>>(0..modulation.len(), 0.0..1.0)?;
 
         chart
             .configure_mesh()
@@ -90,10 +89,10 @@ impl PlottersBackend {
 
     fn plot_1d_impl<B: plotters::backend::DrawingBackend>(
         root: &DrawingArea<B, Shift>,
-        observe_points: &[f64],
+        observe_points: &[f32],
         acoustic_pressures: &[autd3_driver::defined::Complex],
         x_label: &str,
-        yrange: (f64, f64),
+        yrange: (f32, f32),
         config: &PlotConfig,
     ) -> Result<(), crate::error::VisualizerError>
     where
@@ -104,7 +103,7 @@ impl PlottersBackend {
 
         let xrange = observe_points
             .iter()
-            .fold((f64::MAX, f64::MIN), |acc, &x| (acc.0.min(x), acc.1.max(x)));
+            .fold((f32::MAX, f32::MIN), |acc, &x| (acc.0.min(x), acc.1.max(x)));
 
         let x_labels = ((xrange.1 - xrange.0).floor() / config.ticks_step) as usize + 1;
 
@@ -141,13 +140,13 @@ impl PlottersBackend {
     #[allow(clippy::too_many_arguments)]
     fn plot_2d_impl<B: plotters::backend::DrawingBackend>(
         root: &DrawingArea<B, Shift>,
-        observe_points_x: &[f64],
-        observe_points_y: &[f64],
+        observe_points_x: &[f32],
+        observe_points_y: &[f32],
         acoustic_pressures: &[autd3_driver::defined::Complex],
         x_label: &str,
         y_label: &str,
-        zrange: (f64, f64),
-        resolution: f64,
+        zrange: (f32, f32),
+        resolution: f32,
         config: &PlotConfig,
     ) -> Result<(), crate::error::VisualizerError>
     where
@@ -156,7 +155,7 @@ impl PlottersBackend {
     {
         root.fill(&WHITE)?;
 
-        let main_area_size_x = (config.figsize.0 as f64 * (1.0 - config.cbar_size)) as u32;
+        let main_area_size_x = (config.figsize.0 as f32 * (1.0 - config.cbar_size)) as u32;
 
         let (main_area, cbar_area) = root.split_horizontally(main_area_size_x);
 
@@ -168,10 +167,10 @@ impl PlottersBackend {
         {
             let xrange = observe_points_x
                 .iter()
-                .fold((f64::MAX, f64::MIN), |acc, &x| (acc.0.min(x), acc.1.max(x)));
+                .fold((f32::MAX, f32::MIN), |acc, &x| (acc.0.min(x), acc.1.max(x)));
             let yrange = observe_points_y
                 .iter()
-                .fold((f64::MAX, f64::MIN), |acc, &x| (acc.0.min(x), acc.1.max(x)));
+                .fold((f32::MAX, f32::MIN), |acc, &x| (acc.0.min(x), acc.1.max(x)));
 
             let plot_range_x = xrange.1 - xrange.0;
             let plot_range_y = yrange.1 - yrange.0;
@@ -182,8 +181,8 @@ impl PlottersBackend {
             let available_size_x = main_area_size_x - config.label_area_size - config.margin;
             let available_size_y = config.figsize.1 - config.label_area_size - config.margin * 2;
 
-            let px_per_ps = (available_size_x as f64 / plot_range_x)
-                .min(available_size_y as f64 / plot_range_y);
+            let px_per_ps = (available_size_x as f32 / plot_range_x)
+                .min(available_size_y as f32 / plot_range_y);
 
             let plot_size_x = (plot_range_x * px_per_ps) as u32;
             let plot_size_y = (plot_range_y * px_per_ps) as u32;
@@ -258,7 +257,7 @@ impl PlottersBackend {
                 .y_label_formatter(&|&v| {
                     format!(
                         "{:.2}",
-                        zrange.0 + (zrange.1 - zrange.0) * v as f64 / color_map_size as f64
+                        zrange.0 + (zrange.1 - zrange.0) * v as f32 / color_map_size as f32
                     )
                 })
                 .y_desc("Amplitude [-]")
@@ -286,7 +285,7 @@ impl PlottersBackend {
         root: DrawingArea<B, Shift>,
         config: &PlotConfig,
         geometry: &Geometry,
-        phases: Vec<f64>,
+        phases: Vec<f32>,
     ) -> Result<(), crate::error::VisualizerError>
     where
         VisualizerError:
@@ -294,7 +293,7 @@ impl PlottersBackend {
     {
         root.fill(&WHITE)?;
 
-        let main_area_size_x = (config.figsize.0 as f64 * (1.0 - config.cbar_size)) as u32;
+        let main_area_size_x = (config.figsize.0 as f32 * (1.0 - config.cbar_size)) as u32;
 
         let (main_area, cbar_area) = root.split_horizontally(main_area_size_x);
 
@@ -310,13 +309,13 @@ impl PlottersBackend {
                 .collect::<Vec<_>>();
 
             let min_x =
-                p.iter().fold(f64::MAX, |acc, &(x, _)| acc.min(x)) - AUTD3::TRANS_SPACING / 2.0;
+                p.iter().fold(f32::MAX, |acc, &(x, _)| acc.min(x)) - AUTD3::TRANS_SPACING / 2.0;
             let min_y =
-                p.iter().fold(f64::MAX, |acc, &(_, y)| acc.min(y)) - AUTD3::TRANS_SPACING / 2.0;
+                p.iter().fold(f32::MAX, |acc, &(_, y)| acc.min(y)) - AUTD3::TRANS_SPACING / 2.0;
             let max_x =
-                p.iter().fold(f64::MIN, |acc, &(x, _)| acc.max(x)) + AUTD3::TRANS_SPACING / 2.0;
+                p.iter().fold(f32::MIN, |acc, &(x, _)| acc.max(x)) + AUTD3::TRANS_SPACING / 2.0;
             let max_y =
-                p.iter().fold(f64::MIN, |acc, &(_, y)| acc.max(y)) + AUTD3::TRANS_SPACING / 2.0;
+                p.iter().fold(f32::MIN, |acc, &(_, y)| acc.max(y)) + AUTD3::TRANS_SPACING / 2.0;
 
             let plot_range_x = max_x - min_x;
             let plot_range_y = max_y - min_y;
@@ -324,8 +323,8 @@ impl PlottersBackend {
             let available_size_x = main_area_size_x - config.label_area_size - config.margin;
             let available_size_y = config.figsize.1 - config.label_area_size - config.margin * 2;
 
-            let px_per_ps = (available_size_x as f64 / plot_range_x)
-                .min(available_size_y as f64 / plot_range_y);
+            let px_per_ps = (available_size_x as f32 / plot_range_x)
+                .min(available_size_y as f32 / plot_range_y);
 
             let plot_size_x = (plot_range_x * px_per_ps) as u32;
             let plot_size_y = (plot_range_y * px_per_ps) as u32;
@@ -365,7 +364,7 @@ impl PlottersBackend {
 
             scatter_ctx.draw_series(p.iter().zip(phases.iter()).map(|(&(x, y), &p)| {
                 let v = (p / (2.0 * autd3_driver::defined::PI)) % 1.;
-                let c = cmap[((v * color_map_size as f64) as usize).clamp(0, cmap.len() - 1)];
+                let c = cmap[((v * color_map_size as f32) as usize).clamp(0, cmap.len() - 1)];
                 Circle::new(
                     (x, y),
                     AUTD3::TRANS_SPACING * px_per_ps / 2.0,
@@ -437,9 +436,9 @@ impl Backend for PlottersBackend {
     }
 
     fn plot_1d(
-        observe_points: Vec<f64>,
+        observe_points: Vec<f32>,
         acoustic_pressures: Vec<autd3_driver::defined::Complex>,
-        _resolution: f64,
+        _resolution: f32,
         x_label: &str,
         config: Self::PlotConfig,
     ) -> Result<(), crate::error::VisualizerError> {
@@ -450,7 +449,7 @@ impl Backend for PlottersBackend {
 
         let yrange = acoustic_pressures
             .iter()
-            .fold((f64::MAX, f64::MIN), |acc, &x| {
+            .fold((f32::MAX, f32::MIN), |acc, &x| {
                 (acc.0.min(x.norm()), acc.1.max(x.norm()))
             });
 
@@ -476,10 +475,10 @@ impl Backend for PlottersBackend {
     }
 
     fn plot_2d(
-        observe_x: Vec<f64>,
-        observe_y: Vec<f64>,
+        observe_x: Vec<f32>,
+        observe_y: Vec<f32>,
         acoustic_pressures: Vec<autd3_driver::defined::Complex>,
-        resolution: f64,
+        resolution: f32,
         x_label: &str,
         y_label: &str,
         config: Self::PlotConfig,
@@ -491,7 +490,7 @@ impl Backend for PlottersBackend {
 
         let zrange = acoustic_pressures
             .iter()
-            .fold((f64::MAX, f64::MIN), |acc, &x| {
+            .fold((f32::MAX, f32::MIN), |acc, &x| {
                 (acc.0.min(x.norm()), acc.1.max(x.norm()))
             });
 
@@ -523,7 +522,7 @@ impl Backend for PlottersBackend {
     }
 
     fn plot_modulation(
-        modulation: Vec<f64>,
+        modulation: Vec<f32>,
         config: Self::PlotConfig,
     ) -> Result<(), crate::error::VisualizerError> {
         let path = std::path::Path::new(&config.fname);
@@ -549,7 +548,7 @@ impl Backend for PlottersBackend {
     fn plot_phase(
         config: Self::PlotConfig,
         geometry: &autd3_driver::geometry::Geometry,
-        phases: Vec<f64>,
+        phases: Vec<f32>,
     ) -> Result<(), crate::error::VisualizerError> {
         let path = std::path::Path::new(&config.fname);
         if !path.parent().map_or(true, |p| p.exists()) {
